wordifier/helper.py
import enchant


# Dictionary object for checking word in english
english_dict = enchant.Dict("en_US")

# Map for number(Digit) to corresponding alphabets
number_to_alpha_dict = {
    "1": [],
    "2": ["A", "B", "C"],
    "3": ["D", "E", "F"],
    "4": ["G", "H", "I"],
    "5": ["J", "K", "L"],
    "6": ["M", "N", "O"],
    "7": ["P", "Q", "R", "S"],
    "8": ["T", "U", "V"],
    "9": ["W", "X", "Y", "Z"],
    "0": []
}
